application_component/python_code/requirements.txt
annotated-types==0.7.0
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
ffmpeg-python==0.2.0
future==1.0.0
gevent==24.2.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
numpy==2.0.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyparsing==3.1.2
PySocks==1.7.1
python-dotenv==1.0.1
pyttsx3==2.90
requests==2.32.3
rsa==4.9
selenium==4.23.1
setuptools==71.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.4
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
